requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-groq
langchain-text-splitters
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
faiss-cpu
python-dotenv
pypdf
pymupdf
bs4
arxiv
wikipedia
validators==0.28.1
youtube-transcript-api
pytube==15.0.0
numexpr
duckduckgo-search
uvicorn
sse-starlette
langserve
fastapi==0.110.0
pydantic>=2.0
mysql-connector-python
langchain-mysql
SQLAlchemy
psycopg2-binary
